calculate interest by ms
