+++ contract/src/common/mod.rs
@@ -14,3 +14,4 @@ pub type TokenAmount = u128;
 pub(crate) const MS_IN_SECOND: u64 = 1000;
 pub(crate) const MS_IN_MINUTE: u64 = MS_IN_SECOND * 60;
 pub(crate) const MINUTES_IN_YEAR: u64 = 365 * 24 * 60;
+pub(crate) const MS_IN_YEAR: Duration = MINUTES_IN_YEAR * MS_IN_MINUTE;
