+++ integration-tests/src/common.rs
@@ -2,6 +2,7 @@ use serde_json::Value;
 
 pub trait ValueGetters {
     fn get_u128(&self, key: &str) -> u128;
+    fn get_interest(&self) -> u128;
 }
 
 impl ValueGetters for Value {
@@ -15,5 +16,9 @@ impl ValueGetters for Value {
             .to_string()
             .parse::<u128>()
             .unwrap()
+    }
+
+    fn get_interest(&self) -> u128 {
+        self.as_object().unwrap().get("amount").unwrap().get_u128("total")
     }
 }
